Fix it_gets_wasm_bindgen_version test
This test case creates a Cargo.toml fixture specifying version "0.3" of
wasm-bindgen-test as a dependency, then reads the lockfile and checks
the version of that dependency therein.  It previously checked for a
specific, hard-coded version of that dependency in the lock file, but
that fails once the published package has been updated.  Instead, we
should check that the version in the lockfile matches the semver
specified in Cargo.toml.